feat: location과 ride 도메인 구조를 정리
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
