--- src/main/java/com/bikeprojectminji/bikeback/dto/ride/RideRecordPointRequest.java
@@ -1,10 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.ride;
-
-import java.math.BigDecimal;
-
-public record RideRecordPointRequest(
-        Integer pointOrder,
-        BigDecimal latitude,
-        BigDecimal longitude
-) {
-}
