--- src/main/java/com/bikeprojectminji/bikeback/dto/ride/RideRecordResponse.java
@@ -1,8 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.ride;
-
-public record RideRecordResponse(
-        Long rideRecordId,
-        Long ownerUserId,
-        Integer routePointCount
-) {
-}
