--- src/main/java/com/bikeprojectminji/bikeback/dto/ride/RideRecordSummaryRequest.java
@@ -1,7 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.ride;
-
-public record RideRecordSummaryRequest(
-        Integer distanceM,
-        Integer durationSec
-) {
-}
