--- src/main/java/com/bikeprojectminji/bikeback/dto/ridepolicy/RidePolicyEvaluationRequest.java
@@ -1,7 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.ridepolicy;
-
-public record RidePolicyEvaluationRequest(
-        String phase,
-        RideLocationRequest location
-) {
-}
