--- src/main/java/com/bikeprojectminji/bikeback/dto/ridepolicy/RidePolicyEvaluationResponse.java
@@ -1,10 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.ridepolicy;
-
-public record RidePolicyEvaluationResponse(
-        String phase,
-        RidePolicyGateResponse startGate,
-        RidePolicyGateResponse offRoute,
-        String overallState,
-        String defaultMessage
-) {
-}
