--- src/main/java/com/bikeprojectminji/bikeback/dto/ridepolicy/RidePolicyGateResponse.java
@@ -1,7 +1,0 @@
-package com.bikeprojectminji.bikeback.dto.ridepolicy;
-
-public record RidePolicyGateResponse(
-        String status,
-        String reasonCode
-) {
-}
