+++ src/main/java/com/bikeprojectminji/bikeback/location/service/RecentLocationCacheStore.java
@@ -1,0 +1,10 @@
+package com.bikeprojectminji.bikeback.location.service;
+
+import java.util.Optional;
+
+public interface RecentLocationCacheStore {
+
+    Optional<RecentLocationSnapshot> find(String subject);
+
+    void save(String subject, RecentLocationSnapshot snapshot);
+}
