+++ src/main/java/com/bikeprojectminji/bikeback/location/service/RecentLocationStatus.java
@@ -1,0 +1,6 @@
+package com.bikeprojectminji.bikeback.location.service;
+
+public enum RecentLocationStatus {
+    ACTIVE,
+    COMPLETE
+}
