+++ src/main/java/com/bikeprojectminji/bikeback/ride/dto/CreateRideRecordRequest.java
@@ -1,0 +1,12 @@
+package com.bikeprojectminji.bikeback.ride.dto;
+
+import java.time.OffsetDateTime;
+import java.util.List;
+
+public record CreateRideRecordRequest(
+        OffsetDateTime startedAt,
+        OffsetDateTime endedAt,
+        RideRecordSummaryRequest summary,
+        List<RideRecordPointRequest> routePoints
+) {
+}
