+++ src/main/java/com/bikeprojectminji/bikeback/ride/dto/RideRecordPointRequest.java
@@ -1,0 +1,10 @@
+package com.bikeprojectminji.bikeback.ride.dto;
+
+import java.math.BigDecimal;
+
+public record RideRecordPointRequest(
+        Integer pointOrder,
+        BigDecimal latitude,
+        BigDecimal longitude
+) {
+}
