+++ src/main/java/com/bikeprojectminji/bikeback/ride/dto/RideRecordResponse.java
@@ -1,0 +1,8 @@
+package com.bikeprojectminji.bikeback.ride.dto;
+
+public record RideRecordResponse(
+        Long rideRecordId,
+        Long ownerUserId,
+        Integer routePointCount
+) {
+}
