+++ src/main/java/com/bikeprojectminji/bikeback/ride/dto/RideRecordSummaryRequest.java
@@ -1,0 +1,7 @@
+package com.bikeprojectminji.bikeback.ride.dto;
+
+public record RideRecordSummaryRequest(
+        Integer distanceM,
+        Integer durationSec
+) {
+}
