+++ src/main/java/com/bikeprojectminji/bikeback/ride/policy/dto/RidePolicyEvaluationRequest.java
@@ -1,0 +1,7 @@
+package com.bikeprojectminji.bikeback.ride.policy.dto;
+
+public record RidePolicyEvaluationRequest(
+        String phase,
+        RideLocationRequest location
+) {
+}
