+++ src/main/java/com/bikeprojectminji/bikeback/ride/policy/dto/RidePolicyEvaluationResponse.java
@@ -1,0 +1,10 @@
+package com.bikeprojectminji.bikeback.ride.policy.dto;
+
+public record RidePolicyEvaluationResponse(
+        String phase,
+        RidePolicyGateResponse startGate,
+        RidePolicyGateResponse offRoute,
+        String overallState,
+        String defaultMessage
+) {
+}
