+++ src/main/java/com/bikeprojectminji/bikeback/ride/policy/dto/RidePolicyGateResponse.java
@@ -1,0 +1,7 @@
+package com.bikeprojectminji.bikeback.ride.policy.dto;
+
+public record RidePolicyGateResponse(
+        String status,
+        String reasonCode
+) {
+}
